Remove Python tooling in favor of Bun
