+++ .github/workflows/pages.yml
@@ -27,21 +27,13 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
 
-      - name: Set up Python
-        uses: actions/setup-python@5fda3b95a4ea91299a34e894583c3862153e4b97 # v7.0.0
-        with:
-          python-version: "3.13.14"
-          cache: pip
-
       - name: Set up Bun
         uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
         with:
           bun-version: "1.3.14"
 
       - name: Install dependencies
-        run: |
-          python -m pip install -r requirements.txt
-          bun install --frozen-lockfile
+        run: bun install --frozen-lockfile
 
       - name: Run tests
         run: bun run test
@@ -118,5 +110,5 @@ jobs:
           response="$(curl --fail-with-body --silent --show-error \
             --retry 5 --retry-all-errors --retry-delay 2 \
             'https://flash.4st.li/api/games?q=bike%20mania')"
-          RESPONSE="$response" python -c \
-            'import json, os; payload = json.loads(os.environ["RESPONSE"]); assert isinstance(payload.get("games"), list)'
+          RESPONSE="$response" bun -e \
+            'const payload = JSON.parse(process.env.RESPONSE); if (!Array.isArray(payload.games)) throw new Error("Invalid catalog response");'
