+++ package.json
@@ -1,12 +1,17 @@
 {
   "packageManager": "bun@1.3.14",
   "scripts": {
-    "test": "bun run typecheck && bun test tests/deploy.test.ts tests/catalog.test.ts tests/dev-server.test.ts && python -m unittest discover -s tests -p 'test_*.py' -v && node --check site/js/games.js && node --check site/js/game-installer.js && node --check site/js/game-library.js && node --check site/js/filesystem.js && node --check site/js/file-operations.js && node --check site/js/dialogs.js && node --check site/js/offline.js && node --check site/js/offline-worker.js && node --check site/js/main.js && node tools/validate_icons.mjs",
+    "test": "bun run typecheck && bun test && bun run validate:javascript && bun run validate:icons",
     "build": "bun tools/deploy.ts",
     "dev": "bun tools/dev-server.ts",
     "deploy:worker": "wrangler deploy --config worker/wrangler.toml",
+    "extract:swf": "bun tools/extract-swf.ts",
+    "find:icons": "bun tools/find-flashpoint-icons.ts",
+    "sync:icons": "bun tools/sync-game-icons.ts",
     "typecheck": "tsc --noEmit",
-    "validate:icons": "node tools/validate_icons.mjs --require-all"
+    "update:ruffle": "bun tools/update-ruffle.ts",
+    "validate:icons": "bun tools/validate-icons.ts --require-all",
+    "validate:javascript": "bun tools/validate-javascript.ts"
   },
   "devDependencies": {
     "@types/bun": "^1.3.14",
@@ -15,6 +20,7 @@
     "wrangler": "4.114.0"
   },
   "dependencies": {
-    "fflate": "0.8.3"
+    "fflate": "0.8.3",
+    "sharp": "^0.35.3"
   }
 }
