--- requirements.txt
@@ -1,2 +1,0 @@
-Pillow>=11.0
-requests>=2.32
