--- tests/offline_node_tests.js
@@ -1,306 +1,0 @@
-"use strict";
-
-const assert = require("assert");
-const {
-  BUNDLED_GAME_CACHE,
-  createManager,
-  validateGameManifest,
-  waitForWorker,
-} = require("../site/js/offline.js");
-
-class Events {
-  constructor() {
-    this.listeners = new Map();
-  }
-  addEventListener(name, listener) {
-    if (!this.listeners.has(name)) this.listeners.set(name, new Set());
-    this.listeners.get(name).add(listener);
-  }
-  removeEventListener(name, listener) {
-    this.listeners.get(name)?.delete(listener);
-  }
-  dispatch(name) {
-    this.listeners.get(name)?.forEach((listener) => listener());
-  }
-}
-
-class Worker extends Events {
-  constructor(state = "activated") {
-    super();
-    this.state = state;
-    this.messages = [];
-  }
-  transition(state) {
-    this.state = state;
-    this.dispatch("statechange");
-  }
-  postMessage(message) {
-    this.messages.push(message);
-  }
-}
-
-class Registration extends Events {
-  constructor({ active = null, installing = null, waiting = null } = {}) {
-    super();
-    this.active = active;
-    this.installing = installing;
-    this.waiting = waiting;
-    this.updateCalls = 0;
-    this.unregisterCalls = 0;
-  }
-  async update() {
-    this.updateCalls += 1;
-  }
-  async unregister() {
-    this.unregisterCalls += 1;
-  }
-}
-
-class MemoryStorage {
-  constructor(values = {}) {
-    this.values = new Map(Object.entries(values));
-  }
-  getItem(key) {
-    return this.values.get(key) ?? null;
-  }
-  setItem(key, value) {
-    this.values.set(key, String(value));
-  }
-}
-
-class MemoryCache {
-  constructor() {
-    this.values = new Map();
-  }
-  async put(key, response) {
-    this.values.set(String(key), response);
-  }
-  async delete(key) {
-    return this.values.delete(String(key));
-  }
-}
-
-const manifest = {
-  version: "26.07.29-aaaaaaa",
-  runtime: {
-    revision: "runtime-1",
-    bytes: 10,
-    files: [{ url: "js/runtime.wasm", bytes: 10 }],
-  },
-  games: {
-    "bike-mania": {
-      revision: "bike-1",
-      type: "swf",
-      bytes: 4,
-      files: [{ url: "swf/bike-mania/main.swf", bytes: 4 }],
-    },
-    doom: {
-      revision: "doom-1",
-      type: "iframe",
-      bytes: 6,
-      files: [
-        { url: "iframe/doom/index.html", bytes: 2 },
-        { url: "dos/doom/doom.jsdos", bytes: 4 },
-      ],
-    },
-  },
-};
-
-const makeEnvironment = ({
-  registration = new Registration({ active: new Worker() }),
-  remoteVersion = "26.07.29-aaaaaaa",
-  storageValues = {},
-} = {}) => {
-  const serviceWorker = new Events();
-  serviceWorker.controller = registration.active;
-  serviceWorker.registerCalls = [];
-  serviceWorker.register = async (...args) => {
-    serviceWorker.registerCalls.push(args);
-    return registration;
-  };
-  serviceWorker.getRegistration = async () => registration;
-  serviceWorker.ready = Promise.resolve(registration);
-  const bundledCache = new MemoryCache();
-  const shellCache = new MemoryCache();
-  const deletedCaches = [];
-  const storage = new MemoryStorage({
-    astroFlashLastUpdateCheck: "1800000000000",
-    ...storageValues,
-  });
-  let reloads = 0;
-  const environment = new Events();
-  Object.assign(environment, {
-    navigator: {
-      onLine: true,
-      serviceWorker,
-      storage: {
-        estimate: async () => ({ usage: 100, quota: 10_000 }),
-      },
-    },
-    localStorage: storage,
-    location: {
-      href: "https://flash.example/",
-      origin: "https://flash.example",
-      reload: () => {
-        reloads += 1;
-      },
-    },
-    Date: { now: () => 1_800_000_000_000 },
-    document: Object.assign(new Events(), { visibilityState: "visible" }),
-    caches: {
-      keys: async () => ["astro-flash-precache", BUNDLED_GAME_CACHE],
-      open: async (name) =>
-        name === BUNDLED_GAME_CACHE ? bundledCache : shellCache,
-      delete: async (name) => {
-        deletedCaches.push(name);
-        if (name === BUNDLED_GAME_CACHE) bundledCache.values.clear();
-        return true;
-      },
-    },
-    fetch: async (url, options) => {
-      assert.strictEqual(options.cache, "no-store");
-      if (String(url).startsWith("version.json")) {
-        return new Response(
-          JSON.stringify({
-            version: remoteVersion,
-            revision: remoteVersion.split("-").at(-1),
-            offlineBytes: 8_000_000,
-            bundledGameBytes: 10,
-          }),
-        );
-      }
-      if (url === "offline-games.json") {
-        return new Response(JSON.stringify(manifest));
-      }
-      return new Response(new Uint8Array([1]));
-    },
-  });
-  return {
-    bundledCache,
-    deletedCaches,
-    environment,
-    getReloads: () => reloads,
-    registration,
-    serviceWorker,
-    storage,
-  };
-};
-
-(async () => {
-  const activatingWorker = new Worker("installing");
-  const activation = waitForWorker(activatingWorker);
-  activatingWorker.transition("activated");
-  await activation;
-
-  assert.throws(
-    () =>
-      validateGameManifest({
-        ...manifest,
-        runtime: {
-          ...manifest.runtime,
-          files: [{ url: "../runtime.wasm", bytes: 1 }],
-        },
-      }),
-    /invalid/,
-  );
-
-  const initial = makeEnvironment();
-  const manager = createManager({
-    currentVersion: manifest.version,
-    environment: initial.environment,
-  });
-  await manager.initialize();
-  assert.deepStrictEqual(initial.serviceWorker.registerCalls[0], [
-    "sw.js",
-    { updateViaCache: "none" },
-  ]);
-  assert.strictEqual(manager.getSnapshot().phase, "ready");
-  assert.strictEqual(manager.getSnapshot().bundledGames.length, 2);
-
-  await manager.downloadGame("bike-mania");
-  assert.deepStrictEqual(manager.getSnapshot().downloadedGameIds, [
-    "bike-mania",
-  ]);
-  assert.strictEqual(manager.getSnapshot().downloadedGameBytes, 14);
-  assert.strictEqual(initial.bundledCache.values.size, 2);
-
-  await manager.downloadGame("doom");
-  assert.strictEqual(manager.getSnapshot().downloadedGameIds.length, 2);
-  assert.strictEqual(manager.getSnapshot().downloadedGameBytes, 20);
-  assert.strictEqual(initial.bundledCache.values.size, 4);
-
-  await manager.removeGame("bike-mania");
-  assert.deepStrictEqual(manager.getSnapshot().downloadedGameIds, ["doom"]);
-  assert.strictEqual(manager.getSnapshot().downloadedGameBytes, 6);
-  assert.strictEqual(
-    initial.bundledCache.values.has("https://flash.example/js/runtime.wasm"),
-    false,
-  );
-
-  await manager.removeAllGames();
-  assert.deepStrictEqual(manager.getSnapshot().downloadedGameIds, []);
-  assert(initial.deletedCaches.includes(BUNDLED_GAME_CACHE));
-
-  const staleRuntime = makeEnvironment({
-    storageValues: {
-      astroFlashOfflineGameRecords: JSON.stringify({
-        __runtime__: {
-          bytes: 10,
-          files: ["js/old-runtime.wasm"],
-          revision: "runtime-old",
-          type: "runtime",
-        },
-        "bike-mania": {
-          bytes: 4,
-          files: ["swf/bike-mania/main.swf"],
-          revision: "bike-1",
-          type: "swf",
-        },
-      }),
-    },
-  });
-  const staleRuntimeManager = createManager({
-    currentVersion: manifest.version,
-    environment: staleRuntime.environment,
-  });
-  await staleRuntimeManager.initialize();
-  for (let attempt = 0; attempt < 10; attempt += 1) {
-    const saved = JSON.parse(
-      staleRuntime.storage.getItem("astroFlashOfflineGameRecords"),
-    );
-    if (saved.__runtime__?.revision === manifest.runtime.revision) break;
-    await new Promise((resolve) => setImmediate(resolve));
-  }
-  assert.strictEqual(
-    JSON.parse(
-      staleRuntime.storage.getItem("astroFlashOfflineGameRecords"),
-    ).__runtime__.revision,
-    manifest.runtime.revision,
-  );
-
-  const updateRegistration = new Registration({ active: new Worker() });
-  const update = makeEnvironment({
-    registration: updateRegistration,
-    remoteVersion: "26.07.30-bbbbbbb",
-  });
-  const updateManager = createManager({
-    currentVersion: manifest.version,
-    environment: update.environment,
-  });
-  await updateManager.initialize();
-  await updateManager.checkForUpdates();
-  assert.strictEqual(updateManager.getSnapshot().availableVersion, "26.07.30-bbbbbbb");
-  const waitingWorker = new Worker("installed");
-  updateRegistration.waiting = waitingWorker;
-  updateRegistration.dispatch("updatefound");
-  await new Promise((resolve) => setImmediate(resolve));
-  await updateManager.applyUpdate();
-  assert.deepStrictEqual(waitingWorker.messages, [{ type: "SKIP_WAITING" }]);
-  update.serviceWorker.dispatch("controllerchange");
-  assert.strictEqual(update.getReloads(), 1);
-
-  console.log("offline update tests passed");
-})().catch((error) => {
-  console.error(error);
-  process.exitCode = 1;
-});
